backend/pyproject.toml
[project]
name = "openforge-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.120.4",
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.0.0",
    "uvicorn[standard]>=0.38.0",
    "mangum>=0.18.0",
]
